Fix compatibility with Python 3.7 and up
This commit works around a requirement introduced in Python 3.7 that all
calls to fork() be preceeded by calls to PyOS_BeforeFork() in the parent
and PyOS_AfterFork_Parent() and PyOS_AfterFork_Child() in the respective
processes.

Also, Python 3.7 always initialize threads, but if the option is not set
on uWSGI the thread-local data is not set and it causes random crashes,
therefore it is now a fatal error if that option is not set.

A plethora of new hooks were added because Python specifies that after
fork the first calls must be to PyOS_AfterFork_* and unfortunately uWSGI
was doing a bunch of stuff before calling the existing post_fork() hook.

We decided to modify the function uwsgi_fork() itself because it would
make it easier to handle all use-cases: mules, spoolers, re-forking of
workers, lazy apps, etc.

Reference:

https://docs.python.org/3.7/c-api/sys.html#c.PyOS_BeforeFork
https://docs.python.org/3.7/c-api/init.html#c.PyEval_InitThreads
https://github.com/unbit/uwsgi/pull/1995
https://github.com/unbit/uwsgi/issues/1978
https://github.com/awelzel/uwsgi/tree/before-after-fork