Disable XML support on build
An uwsgi built with XML support dynamically loads the system's libxml
library when ran, but that version is not compatible with the versions
our XML python libs expect. Because the lib is already loaded, when the
python libs are imported they skip loading their own static versions of
libxml and instead try to link up with the version loaded by uwsgi,
which obviously results in an error.

Since we don't use XML in our uwsgi configs, we'll just disable XML
support.